Added new page template "Single Column Narrow"
Added new page template "Single Column Narrow", which uses a single
column layout with the secondary sidebar displayed below instead of to
the right. The maximum content with would be 60em for this template,
designed for optimum line length. This was previously the behaviour of
the default page template. The default page template now uses the same
behaviour as the posts and archive pages, of displaying the secondary
sidebar to the right if active.